# tests/fixtures/application_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SlideFlow",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tasks": {
      "get": {
        "summary": "List Tasks",
        "description": "获取任务列表，支持 status / task_type / owner_id 过滤、追踪码搜索及分页。\n\n权限规则：\n- 管理员：可传任意 owner_id（不传则返回所有用户的任务）\n- 非管理员：忽略 owner_id 参数，强制只看自己的任务\n\n分页：page 从 1 开始，page_size 默认 20、上限 100。\n搜索：search 仅在 task_type=download 时对 params.track_code 字段做模糊匹配。",
        "operationId": "list_tasks_api_tasks_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "task_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Type"
            }
          },
          {
            "name": "owner_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Tasks Api Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}": {
      "get": {
        "summary": "Get Task",
        "description": "获取单个任务详情。只能查看自己的任务（管理员可查看所有）",
        "operationId": "get_task_api_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Task Api Tasks  Task Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/split-import": {
      "post": {
        "summary": "Create Split Import Task",
        "description": "异步批量拆分导入：创建任务记录后立即返回，后台线程执行拆分",
        "operationId": "create_split_import_task_api_tasks_split_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_split_import_task_api_tasks_split_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Split Import Task Api Tasks Split Import Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/cancel": {
      "post": {
        "summary": "Cancel Task",
        "description": "取消正在执行的任务",
        "operationId": "cancel_task_api_tasks__task_id__cancel_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cancel Task Api Tasks  Task Id  Cancel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/tasks/bulk-delete": {
      "post": {
        "summary": "Bulk Delete Tasks",
        "description": "批量删除任务记录，并清理任务遗留的临时上传目录。",
        "operationId": "bulk_delete_tasks_api_admin_tasks_bulk_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskDeletePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Bulk Delete Tasks Api Admin Tasks Bulk Delete Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/tasks/{task_id}": {
      "delete": {
        "summary": "Delete Task",
        "description": "删除单个任务记录，并清理任务遗留的临时上传目录。",
        "operationId": "delete_task_api_admin_tasks__task_id__delete",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Task Api Admin Tasks  Task Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_split_import_task_api_tasks_split_import_post": {
        "properties": {
          "name_prefix": {
            "type": "string",
            "title": "Name Prefix"
          },
          "subject": {
            "type": "string",
            "title": "Subject",
            "default": ""
          },
          "tags": {
            "type": "string",
            "title": "Tags",
            "default": ""
          },
          "secrecy_level": {
            "type": "string",
            "title": "Secrecy Level",
            "default": "public"
          },
          "visibility_scope": {
            "type": "string",
            "title": "Visibility Scope",
            "default": ""
          },
          "visible_user_ids": {
            "type": "string",
            "title": "Visible User Ids",
            "default": ""
          },
          "visible_user_tags": {
            "type": "string",
            "title": "Visible User Tags",
            "default": ""
          },
          "management_scope": {
            "type": "string",
            "title": "Management Scope",
            "default": ""
          },
          "manage_user_ids": {
            "type": "string",
            "title": "Manage User Ids",
            "default": ""
          },
          "manage_user_tags": {
            "type": "string",
            "title": "Manage User Tags",
            "default": ""
          },
          "remark_html": {
            "type": "string",
            "title": "Remark Html",
            "default": ""
          },
          "ppt_file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Ppt File"
          },
          "images": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Images"
          }
        },
        "type": "object",
        "required": [
          "name_prefix",
          "ppt_file",
          "images"
        ],
        "title": "Body_create_split_import_task_api_tasks_split_import_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskDeletePayload": {
        "properties": {
          "task_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Task Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "task_ids"
        ],
        "title": "TaskDeletePayload"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}